pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multifuse"
version = "0.1.0"
description = "MultiFuse: Multimodal Weighted Consensus for Action Recognition in Videos"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Nico Grassetto"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "azure-ml-core",
    "azureml-sdk",
    "azure-ai-ml",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "opencv-python>=4.5.0",
    "Pillow>=8.3.0",
    "matplotlib>=3.4.0",
    "seaborn>=0.11.0",
    "tensorboard>=2.7.0",
    "tqdm>=4.62.0",
    "pyyaml>=5.4.0",
    "scikit-learn>=1.0.0",
    "hydra-core>=1.1.0",
]

[project.optional-dependencies]
dev = ["pytest>=6.2.0", "black>=21.7.0", "flake8>=3.9.0"]
notebook = ["jupyter>=1.0.0", "ipywidgets"]
wandb = ["wandb>=0.12.0"]

[project.urls]
Homepage = "https://github.com/NicoGrassetto/multifuse"
Repository = "https://github.com/NicoGrassetto/multifuse"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
